requirements.txt
NHentai-API